% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_lipd.R
\name{batch_files}
\alias{batch_files}
\title{Batch lipd import}
\usage{
batch_files(x)
}
\arguments{
\item{x}{character. Directory that contains .lipd files.}
}
\value{
One list of data frames for each lipd file
}
\description{
Reads in all .lipd files from the given directory, and loads them into a data frame
}

