% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create.range}
\alias{create.range}
\title{Create the range for ensemble table "number" field}
\usage{
create.range(start, len)
}
\arguments{
\item{start}{Number to start at}

\item{len}{Amount of times to loop}
}
\value{
l A vectory of column ints
}
\description{
Create the range for ensemble table "number" field
}

