% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_lipd.R
\name{export_lipd}
\alias{export_lipd}
\title{Batch export to LiPD}
\usage{
export_lipd(d)
}
\arguments{
\item{d}{character, A list of elements. Each element is one file.}
}
\value{
none
}
\description{
Iterate over all the items in a large list of files that are loaded into the global environment
}

